+++ lib/impossible_angle.js
